Create schoolManagingSystem.py
A school managing system I created on a road trip to a camping trip. It is possible to create a school object, add a curriculum with classes objects. Withing the classes in the curriculum you can add student objects with a grade assigned to that specific class. The curriculum (classes) all have a set period, teacher, and type of class like math or science. 
You can get total classes in a school, a students total grade average, and even how many students are there per classroom.